requirements.txt
httplib2==0.9.2
numpy==1.9.3
picamera==1.10
Pillow==3.0.0
wheel==0.24.0
zbar==0.10
